% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{ComputeColumnStatistics}
\alias{ComputeColumnStatistics}
\title{Compute statistics for all columns (i.e. samples or cells)}
\usage{
ComputeColumnStatistics(data, verbose = TRUE)
}
\arguments{
\item{data}{a grandR object}

\item{verbose}{output status messages}
}
\value{
a new grandR object containing additional columns in the \link{Coldata} table:
\itemize{
  \item{p.conv.X: the T-to-C mismatch frequency in the given ("X") subread category}
  \item{percent.new: new overall percentage of new RNA}
  \item{total.reads: the total number of reads (or UMIs, if UMIs were sequences)}
  \item{total.genes: the total number of genes detected}
  \item{percentage per type: the percentage (up to 100!) of the counts of each type in the GeneInfo}
}
}
\description{
Compute statistics for all columns (i.e. samples or cells)
}
\concept{data}
