% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{CreateConvolutionTable}
\alias{CreateConvolutionTable}
\title{Create Convolution Table from a Seurat object}
\usage{
CreateConvolutionTable(data, n.neighbors = 20)
}
\arguments{
\item{data}{a Seurat object}

\item{n.neighbors}{the number of neighbors to be convoluted}
}
\value{
a table with two columns "Cell" and "Pseudobulk"
}
\description{
Create Convolution Table from a Seurat object
}
\details{
This function returns a table which can be used as input for GRAND3. Note that a data set contatining multiple time points should be split before convolution.
}
\concept{data}
