% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{CreatePseudobulkTable}
\alias{CreatePseudobulkTable}
\title{Create Pseudobulk Table from a Seurat object}
\usage{
CreatePseudobulkTable(
  data,
  name.column = "Name",
  pseudobulk.column = "Condition"
)
}
\arguments{
\item{data}{a Seurat object}

\item{name.column}{name of the metadata column containing the sample/cell names. Default "Name".}

\item{pseudobulk.column}{name of the metadata column containing the Pseudobulk names. Default "Condition".}
}
\value{
a table with two columns "Cell" and "Pseudobulk"
}
\description{
Create Pseudobulk Table from a Seurat object
}
\details{
This function returns a table which can be used as input for GRAND3
}
