% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandR.R
\name{data.apply}
\alias{data.apply}
\title{Internal function to apply functions to all slots etc.}
\usage{
data.apply(data, fun, fun.gene.info = NULL, fun.coldata = NULL, ...)
}
\arguments{
\item{data}{a grandR object}

\item{fun}{apply this function to each data slot (i.e. it receives each data matrix)}

\item{fun.gene.info}{apply this function to the gene info table}

\item{fun.coldata}{apply this function to the column annotation table}

\item{...}{passed further to fun, fun.gene.info and fun.coldata}
}
\value{
A new grandR object
}
\description{
Internal function to apply functions to all slots etc.
}
\details{
The additional parameters are provided to each of the functions.
}
\concept{helper}
