% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{density2d}
\alias{density2d}
\title{Density estimation in 2d}
\usage{
density2d(x, y, facet = NULL, n = 100, margin = "n")
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{facet}{factor: estimate for each unique factor; can be NULL}

\item{n}{size of the grid}

\item{margin}{one of 'n','x' or 'y'; should the density be computed along both axes ('n'), or along 'x' or 'y' axis only}
}
\value{
a density value for each point
}
\description{
Estimate point densities on a regular grid for.
}
\concept{helper}
