% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\docType{data}
\name{Design}
\alias{Design}
\title{A list of predefined names for design vectors}
\format{
An object of class \code{list} of length 11.
}
\usage{
Design
}
\description{
These predefined names mainly are implemented here to harmonize analyses.
It is good practise to use these names if sensible.
}
\keyword{datasets}
