% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{FitKineticsPulseR}
\alias{FitKineticsPulseR}
\title{Fit kinetics using pulseR}
\usage{
FitKineticsPulseR(data, name = "pulseR", time = Design$dur.4sU)
}
\arguments{
\item{data}{A grandR object}

\item{name}{the user defined analysis name to store the results}

\item{time}{The column in the column annotation table representing the labeling duration}
}
\value{
a new grandR object containing the pulseR analyses in a new analysis table
}
\description{
Fit kinetics using pulseR
}
\details{
This is adapted code from https://github.com/dieterich-lab/ComparisonOfMetabolicLabeling
}
\concept{kinetics}
