% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\docType{data}
\name{GeneType}
\alias{GeneType}
\title{GeneType}
\format{
An object of class \code{list} of length 4.
}
\usage{
GeneType
}
\description{
A list of functions. These are called one after the other by \link{ReadGRAND} to build the \emph{Type} column. The name of the first function returning TRUE is used as type.
}
\keyword{datasets}
