% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{GetDiagnosticParameters}
\alias{GetDiagnosticParameters}
\title{Describe parameters relevant to diagnostics}
\usage{
GetDiagnosticParameters(data)
}
\arguments{
\item{data}{a grandR object}
}
\value{
a list with
\itemize{
 \item{orientation: Sense or Antisense, only relevant to mismatches for strand unspecific data}
 \item{category: all available categories (Exonic/Intronic, genomes). Note that this might differ from what is available from GeneInfo(data,"Category"), since Grand3 might not have estimated NTRs for all categories!}
 \item{label: which nucleoside analogs have been used}
 \item{model: which model (binom or tbbinom) to inspect}
 \item{estimator: which estimator (joint or separate NTRs were estimated for subreads)}
}
}
\description{
Many of the diagnostics functions expect (optional or mandatory) parameters that are described by this function
}
\concept{diagnostics}
