% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffexp.R
\name{ListGeneSets}
\alias{ListGeneSets}
\title{List available gene sets}
\usage{
ListGeneSets()
}
\value{
the gene set table; use the values in the category and subcategory columns for the corresponding parameters of \link{AnalyzeGeneSets}
}
\description{
Helper function to return a table with all available gene sets for \link{AnalyzeGeneSets}.
}
\details{
This is a convenience wrapper for \link[msigdbr]{msigdbr_collections}.
}
\seealso{
\link{AnalyzeGeneSets}
}
\concept{genesets}
