% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{MAPlot}
\alias{MAPlot}
\title{Make an MA plot}
\usage{
MAPlot(
  data,
  analysis = Analyses(data)[1],
  p.cutoff = 0.05,
  lfc.cutoff = 1,
  annotate.numbers = TRUE,
  ...
)
}
\arguments{
\item{data}{the grandR object that contains the data to be plotted}

\item{analysis}{the analysis to plot (default: first analysis)}

\item{p.cutoff}{p-value cutoff (default: 0.05)}

\item{lfc.cutoff}{log fold change cutoff (default: 1)}

\item{annotate.numbers}{if TRUE, label the number of genes}

\item{...}{further parameters passed to \link{PlotScatter}}
}
\value{
a ggplot object
}
\description{
Plot average expression vs. log2 fold changes
}
\concept{globalplot}
