% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotAnalyses}
\alias{PlotAnalyses}
\title{Convenience function to make the same type of plot for multple analyses.}
\usage{
PlotAnalyses(data, plot.fun, analyses = Analyses(data), add = NULL, ...)
}
\arguments{
\item{data}{the grandR object that contains the data to be plotted}

\item{plot.fun}{the plottinf function to apply}

\item{analyses}{the analyses to plot (default: all)}

\item{add}{additional ggplot (e.g., geoms) objects to add}

\item{...}{passed further to plot.fun}
}
\value{
ggplot objects
}
\description{
Convenience function to make the same type of plot for multple analyses.
}
\concept{globalplot}
