% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{PlotModelCompareErrPrior}
\alias{PlotModelCompareErrPrior}
\title{Diagnostic plot for estimated models (global error rate)}
\usage{
PlotModelCompareErrPrior(
  data,
  label = "4sU",
  estimator = "Separate",
  model = "Binom"
)
}
\arguments{
\item{data}{a grandR object}

\item{label}{which label to consider (see \link{GetDiagnosticParameters}); cannot be NULL}

\item{estimator}{which estimator to consider (see \link{GetDiagnosticParameters}); cannot be NULL}

\item{model}{which model to consider (see \link{GetDiagnosticParameters}); cannot be NULL}
}
\value{
a list with a ggplot object, a description, and the desired size for the plot
}
\description{
This belongs to the fourth kind (model comparison) of diagnostic plots
}
\details{
Compares the prior error rate (estimated from no4sU samples or from all other mismatch types)
against the final error rate estimate.
}
\concept{diagnostics}
