% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{PlotModelCompareNtr}
\alias{PlotModelCompareNtr}
\title{Diagnostic plot for estimated models (global NTR)}
\usage{
PlotModelCompareNtr(data, label = "4sU", estimator = "Separate")
}
\arguments{
\item{data}{a grandR object}

\item{label}{which label to consider (see \link{GetDiagnosticParameters}); cannot be NULL}

\item{estimator}{which estimator to consider (see \link{GetDiagnosticParameters}); cannot be NULL}
}
\value{
a list with a ggplot object, a description, and the desired size for the plot
}
\description{
This belongs to the fourth kind (model comparison) of diagnostic plots
}
\details{
Compares the global NTR (i.e. for all reads used for estimation of
global parameters, what is the percentage of new RNA) for the binom and tbbinom models.
}
\concept{diagnostics}
