% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{PlotModelErr}
\alias{PlotModelErr}
\title{Diagnostic plot for estimated models (global error rate)}
\usage{
PlotModelErr(data, label = "4sU", estimator = "Separate", model = "Binom")
}
\arguments{
\item{data}{a grandR object}

\item{label}{which label to consider (see \link{GetDiagnosticParameters}); cannot be NULL}

\item{estimator}{which estimator to consider (see \link{GetDiagnosticParameters}); cannot be NULL}

\item{model}{which model to consider (see \link{GetDiagnosticParameters}); cannot be NULL}
}
\value{
a list with a ggplot object, a description, and the desired size for the plot
}
\description{
This belongs to the third kind (model) of diagnostic plots
}
\details{
Shows the estimated error rate (i.e., the probability for a conversion
on an old RNA molecule) for each sample.
}
\concept{diagnostics}
