% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{PlotModelLabelTimeCourse}
\alias{PlotModelLabelTimeCourse}
\title{Diagnostic plot for estimated models (4sU increase)}
\usage{
PlotModelLabelTimeCourse(data, label = "4sU", estimator = "Separate")
}
\arguments{
\item{data}{a grandR object}

\item{label}{which label to consider (see \link{GetDiagnosticParameters}); cannot be NULL}

\item{estimator}{which estimator to consider (see \link{GetDiagnosticParameters}); cannot be NULL}
}
\value{
a list with a ggplot object, a description, and the desired size for the plot
}
\description{
This belongs to the third kind (model) of diagnostic plots
}
\details{
Shows the estimated time evolution of 4sU increase in the tbbinom model for each sample.
}
\concept{diagnostics}
