% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{PlotModelShape}
\alias{PlotModelShape}
\title{Diagnostic plot for estimated models (global shape parameter)}
\usage{
PlotModelShape(data, label = "4sU", estimator = "Separate")
}
\arguments{
\item{data}{a grandR object}

\item{label}{which label to consider (see \link{GetDiagnosticParameters}); cannot be NULL}

\item{estimator}{which estimator to consider (see \link{GetDiagnosticParameters}); cannot be NULL}
}
\value{
a list with a ggplot object, a description, and the desired size for the plot
}
\description{
This belongs to the third kind (model) of diagnostic plots
}
\details{
Shows the estimated shape parameter (describing the increase of 4sU over time) in the tbbinom model for each sample.
}
\concept{diagnostics}
