% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{PlotProfileLikelihood}
\alias{PlotProfileLikelihood}
\title{Diagnostic plot for estimated models (global error rate)}
\usage{
PlotProfileLikelihood(
  data,
  label = "4sU",
  estimator = NULL,
  sample = NULL,
  subread = NULL
)
}
\arguments{
\item{data}{a grandR object}

\item{label}{which label to consider (see \link{GetDiagnosticParameters}); cannot be NULL}

\item{estimator}{which estimator to consider (see \link{GetDiagnosticParameters}); cannot be NULL}

\item{sample}{which sample to consider (see \link{GetDiagnosticParameters}); cannot be NULL}

\item{subread}{which subread to consider (see \link{GetDiagnosticParameters}); cannot be NULL}
}
\value{
a list with a ggplot object, a description, and the desired size for the plot
}
\description{
This belongs to the fifth kind (profile likelihoods) of diagnostic plots
}
\details{
Shows the profile likelihoods for all parameters of the tbbinom model.
}
\concept{diagnostics}
