% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotTypeDistribution}
\alias{PlotTypeDistribution}
\title{Plot the distribution of gene types}
\usage{
PlotTypeDistribution(data, mode.slot = DefaultSlot(data), relative = FALSE)
}
\arguments{
\item{data}{the grandR object to get the data to be plotted from}

\item{mode.slot}{which mode and slot to use}

\item{relative}{show percentage values?}
}
\value{
a ggplot object
}
\description{
Plot the distribution of gene types
}
\concept{globalplot}
