% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{RotatateAxisLabels}
\alias{RotatateAxisLabels}
\title{Rotate x axis labels}
\usage{
RotatateAxisLabels(angle = 90)
}
\arguments{
\item{angle}{the angle by which to rotate}
}
\value{
a ggplot theme object
}
\description{
Add this to a ggplot object to rotate the x axis labels
}
\concept{helper}
