% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{Semantics.concentration}
\alias{Semantics.concentration}
\title{Semantics for concentration columns}
\usage{
Semantics.concentration(s, name)
}
\arguments{
\item{s}{original column}

\item{name}{the column name}
}
\value{
a data frame with a single numeric column, where <x>uM from s is replaced by x, <x>mM is replaced by
x*1000, and no4sU is replaced by 0
}
\description{
Defines additional semantics for columns representing concentrations
}
\concept{load}
