% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{Semantics.time}
\alias{Semantics.time}
\title{Semantics for time columns}
\usage{
Semantics.time(s, name)
}
\arguments{
\item{s}{original column}

\item{name}{the column name}
}
\value{
a data frame with a single numeric column, where <x>h from s is replaced by x, <x>min is replaced by
x/60, and no4sU is replaced by 0
}
\description{
Defines additional semantics for columns representing temporal dimensions
}
\concept{load}
