% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{ServeGrandR}
\alias{ServeGrandR}
\title{Serve a shiny web interface}
\usage{
ServeGrandR(
  data,
  table = NULL,
  sizes = NA,
  height = 400,
  plot.gene = NULL,
  plot.global = NULL,
  df.identifier = "Symbol",
  title = Title(data),
  show.sessionInfo = FALSE,
  help = list(".Q: multiple testing corrected p values", ".LFC: log2 fold changes")
)
}
\arguments{
\item{data}{the grandR object (or a file name to an rds file containing a grandR object)}

\item{table}{the table to display (can be NULL; see details)}

\item{sizes}{the widths for the gene plots to show (12 is full screen with); must be a vector as long as there are gene plots}

\item{height}{the height for the gene plots in pixel}

\item{plot.gene}{a list of gene plots; can be NULL, then the stored gene plots are used (see \link{Plots})}

\item{plot.global}{a list of global plots; can be NULL, then the stored global plots are used (see \link{Plots})}

\item{df.identifier}{the main identifier (column name) from the table; this is used when calling the gene plot functions;}

\item{title}{the title to show in the header of the website}

\item{show.sessionInfo}{whether to show session info}

\item{help}{a list of characters that is shown as help text at the beginning (when no gene plot is shown); should describe the contents of your table}
}
\value{
a shiny web server
}
\description{
Fire up a shiny web server for exploratory analysis of grandR data.
}
\details{
If the table parameter is NULL, either an analysis table named "ServeGrandR" is
used (if it exists), otherwise the columns "Q", "LFC", "Synthesis" and "Half-life" of all analysis tables are used.

The gene plots must be functions that accept two parameters: the grandR object and a gene identifier. You can either use
functions directly (e.g. \code{plot.gene=list(PlotGeneOldVsNew)}), or use \link{Defer} in cases you need to specify additional parameters,
e.g. \code{plot.gene=list(Defer(PlotGeneOldVsNew,log=FALSE))}. The global plots are functions accepting a single parameter (the grandR object). Here
the use of \link{Defer} is encouraged due to its caching mechanism.
}
\concept{shiny}
