% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SetParallel}
\alias{SetParallel}
\title{Set up parallel execution}
\usage{
SetParallel(cores = max(1, parallel::detectCores() - 2))
}
\arguments{
\item{cores}{number of cores}
}
\value{
No return value, called for side effects
}
\description{
Set the number of cores for parallel execution.
}
\details{
Whenever \link{psapply} or \link{plapply} are used, they are executed in parallel.
}
\concept{helper}
