% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Summarize}
\alias{Summarize}
\title{Summarize a data matrix}
\usage{
Summarize(mat, summarize.mat)
}
\arguments{
\item{mat}{the matrix to summarize}

\item{summarize.mat}{the matrix defining how to summarize (see details)}
}
\value{
the summarized matrix
}
\description{
Helper function to work in conjunction with \link{GetMatrix} or similar to obtain a summarized matrix.
}
\details{
The summarize.mat can be obtained via \link{GetSummarizeMatrix}. If there are missing (NA) values in the matrix, they are imputed from the rest (average)
}
\concept{data}
