% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grandR.R
\name{UseNtrSlot}
\alias{UseNtrSlot}
\title{Copy the NTR slot and save under new name}
\usage{
UseNtrSlot(data, name)
}
\arguments{
\item{data}{the grandR object}

\item{name}{the name of the new slot}
}
\value{
a grandR object
}
\description{
Copy the NTR slot and save under new name
}
\concept{data}
