% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\docType{data}
\name{Design.Semantics}
\alias{Design.Semantics}
\title{Design.Semantics}
\format{
An object of class \code{list} of length 1.
}
\usage{
Design.Semantics
}
\description{
A list of functions. A function is called by \link{MakeColdata} for each design vector element matching a name of this list.
}
\keyword{datasets}
