% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\docType{data}
\name{GeneType}
\alias{GeneType}
\title{Build the type column for the gene info table.}
\format{
An object of class \code{list} of length 3.
}
\usage{
GeneType
}
\description{
A list of functions. These are called one after the other by \code{\link{ReadGRAND}} or \code{\link{ReadGRAND3}} to build the \emph{Type} column in the \code{\link{GeneInfo}} table. The name of the first function returning TRUE for a row in the table is used as its \emph{type}.
}
\details{
This is by default given to the \emph{classify.genes} parameter of \code{\link{ReadGRAND}} and \code{\link{ReadGRAND3}}.
It will assign the type \emph{mito} if the gene symbol starts with \emph{mt-}, \emph{ERCC} if it starts with \emph{ERCC} and \emph{Cellular} if the gene name starts with \emph{ENS}.
}
\examples{
classi <- c(GeneType,
           `SARS-CoV-2`=function(gene.info) gene.info$Symbol \%in\%
                    c('ORF3a','E','M','ORF6','ORF7a','ORF7b','ORF8','N','ORF10','ORF1ab','S')
           )
sars <- ReadGRAND(system.file("extdata", "sars.tsv.gz", package = "grandR"),
                  design=c("Cell",Design$dur.4sU,Design$Replicate),
                  classify.genes=classi,
                  verbose=TRUE)
table(GeneInfo(sars)$Type)

}
\seealso{
\link{ReadGRAND}, \link{ReadGRAND3}
}
\keyword{datasets}
