% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{MAPlot}
\alias{MAPlot}
\title{Make an MA plot}
\usage{
MAPlot(
  data,
  analysis = Analyses(data)[1],
  aest = aes(),
  p.cutoff = 0.05,
  lfc.cutoff = 1,
  label.numbers = TRUE,
  highlight = NULL,
  label = NULL,
  label.repel = 1
)
}
\arguments{
\item{data}{the grandR object that contains the data to be plotted}

\item{analysis}{the analysis to plot (default: first analysis)}

\item{aest}{parameter to set visual attributes of the plot}

\item{p.cutoff}{p-value cutoff (default: 0.05)}

\item{lfc.cutoff}{log fold change cutoff (default: 1)}

\item{label.numbers}{if TRUE, label the number of genes}

\item{highlight}{highlight these genes; can be either numeric indices, gene names, gene symbols or a logical vector (see details)}

\item{label}{label these genes; can be either numeric indices, gene names, gene symbols or a logical vector (see details)}

\item{label.repel}{force to repel labels from points and each other (increase if labels overlap)}
}
\value{
a ggplot object
}
\description{
Plot average expression vs. log2 fold changes
}
\concept{globalplot}
