% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{psapply}
\alias{psapply}
\alias{plapply}
\title{Parallel (s/l)apply}
\usage{
psapply(..., seed = NULL, enforce = NA)

plapply(..., seed = NULL, enforce = NA)
}
\arguments{
\item{...}{forwarded to lapply or parallel::mclapply}

\item{seed}{Seed for the random number generator}

\item{enforce}{if TRUE, do it parallelized no matter what IsParallel() says, if FALSE do it non-parallelized no matter what IsParallel() says}
}
\value{
a vector (psapply) or list (plapply)
}
\description{
Depending on whether \link{SetParallel} has been called, execute in parallel or not.
}
\details{
If the code uses random number specify the seed to make it deterministic
}
\concept{helper}
